Moved resource functions to GeneralComposables since they are not extension functions.
